+++ tests/operators/test_belnet.py
@@ -2,8 +2,8 @@ import torch
 import pytest
 
 from continuity.operators import BelNet
+from continuity.benchmarks.sine import SineBenchmark
 from continuity.data import OperatorDataset
-from continuity.data.sine import OperatorDataset, Sine
 from continuity.trainer import Trainer
 from continuity.operators.losses import MSELoss
 
@@ -43,17 +43,14 @@ def test_belnet_shape():
 
 @pytest.mark.slow
 def test_belnet():
-    # Parameters
-    num_sensors = 16
-
     # Data set
-    dataset = Sine(num_sensors, size=1)
+    dataset = SineBenchmark(n_train=1).train_dataset
 
     # Operator
     operator = BelNet(dataset.shapes)
 
     # Train
-    Trainer(operator).fit(dataset, tol=1e-3, batch_size=1)
+    Trainer(operator).fit(dataset, tol=1e-3)
 
     # Check solution
     x, u = dataset.x, dataset.u