+++ CHANGELOG.md
@@ -13,6 +13,8 @@
 - Add `timeseries.ipynb` example.
 - Add `Function`, `FunctionSet`, and `FunctionOperatorDataset` classes.
 - Add `function.ipynb` example.
+- Add `Benchmark` base class.
+- Add `SineBenchmark`.
 
 ## 0.0.0 (2024-02-22)